still drafting astrorad (misc. fixes)
git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@3868 74e165f3-c697-40a6-91af-552c8936d7fe

+++ zvtm-cluster-fits/trunk/src/main/java/fr/inria/zvtm/cluster/WallCursor.java
@@ -43,7 +43,7 @@ class WallCursor {
         target.removeGlyph(vRect);
     }
 
-    void moveTo(double x, long y){
+    void moveTo(double x, double y){
         hRect.moveTo(x, y);
         vRect.moveTo(x, y);
     }
